refactor: rewrite cardList and formCardList with hooks
