+++ src/components/CardList/index.tsx
@@ -1,30 +1,25 @@
-import React, { Component } from 'react';
+import React, { FC } from 'react';
 import CardListItem from './CardListItem';
-import { IFlower } from '../../interfaces/flower';
 import { IFlowers } from 'interfaces/flowers';
 
 import './cardList.scss';
 
-class CardList extends Component<IFlowers> {
-  render() {
-    const { flowers } = this.props;
-
-    return (
-      <div className="cards__list">
-        {flowers.map((flower: IFlower) => {
-          return (
-            <CardListItem
-              key={flower.id}
-              id={flower.id}
-              name={flower.name}
-              latinName={flower.latinName}
-              image={flower.image}
-            />
-          );
-        })}
-      </div>
-    );
-  }
-}
+const CardList: FC<IFlowers> = ({ flowers }) => {
+  return (
+    <div className="cards__list">
+      {flowers.map((flower) => {
+        return (
+          <CardListItem
+            key={flower.id}
+            id={flower.id}
+            name={flower.name}
+            latinName={flower.latinName}
+            image={flower.image}
+          />
+        );
+      })}
+    </div>
+  );
+};
 
 export default CardList;
