--- src/interfaces/flower.ts
@@ -1,6 +1,0 @@
-export interface IFlower {
-  id: number;
-  name: string;
-  latinName: string;
-  image: string;
-}
