+++ src/interfaces/flowers.ts
@@ -1,4 +1,9 @@
-import { IFlower } from './flower';
+export interface IFlower {
+  id: number;
+  name: string;
+  latinName: string;
+  image: string;
+}
 
 export interface IFlowers {
   flowers: IFlower[];
