--- src/interfaces/formDataState.ts
@@ -1,5 +1,0 @@
-import { ICard } from './card';
-
-export interface IFormDataState {
-  cards: ICard[];
-}
